Refactor functions
Instead of sourcing all functions at once, only source the phpvm
function. This then passes along all args to a executable.

The advantage here is set -e. As I understand it is running from a
script context, not the shell itself. So when en error occurs the shell
does not exit like it would in a purely sourced function.